+++ packages/common/test/content/_internal/getDownloadsSection.test.ts
@@ -21,9 +21,16 @@ describe("getDownloadsSection", () => {
           scope: "/properties/serverExtractCapability",
           type: "Control",
           options: {
-            helperText: {
-              labelKey: "some.scope.fields.serverExtractCapability.helperText",
-            },
+            control: "hub-field-input-tile-select",
+            labels: [
+              `{{some.scope.fields.serverExtractCapability.exportDataSetting.label:translate}}`,
+              `{{some.scope.fields.serverExtractCapability.defaultDownloadsSystem.label:translate}}`,
+            ],
+            descriptions: [
+              `{{some.scope.fields.serverExtractCapability.exportDataSetting.description:translate}}`,
+              `{{some.scope.fields.serverExtractCapability.defaultDownloadsSystem.description:translate}}`,
+            ],
+            layout: "vertical",
             messages: [
               {
                 type: "CUSTOM",